requirements.txt
paho-mqtt>=1.2.0 # Eclipse Public License v1.0
